+++ python/core/auto_generated/annotations/qgsannotationlinestringitem.sip.in
@@ -1,0 +1,71 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/annotations/qgsannotationlinestringitem.h                   *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+
+class QgsAnnotationLineStringItem : QgsAnnotationItem
+{
+%Docstring
+An annotation item which renders a line symbol along a linestring geometry.
+
+.. versionadded:: 3.16
+%End
+
+%TypeHeaderCode
+#include "qgsannotationlinestringitem.h"
+%End
+  public:
+
+    QgsAnnotationLineStringItem( const QgsLineString &linestring, const QgsCoordinateReferenceSystem &crs );
+%Docstring
+Constructor for QgsAnnotationLineStringItem, with the specified ``linestring`` in the given ``crs``.
+%End
+    ~QgsAnnotationLineStringItem();
+
+    virtual QString type() const;
+
+    virtual void render( QgsRenderContext &context, QgsFeedback *feedback );
+
+    virtual bool writeXml( QDomElement &element, QDomDocument &document, const QgsReadWriteContext &context ) const;
+
+    static QgsAnnotationLineStringItem *create() /Factory/;
+    virtual bool readXml( const QDomElement &element, const QgsReadWriteContext &context );
+
+
+    virtual QgsAnnotationLineStringItem *clone() /Factory/;
+
+
+    const QgsLineSymbol *symbol() const;
+%Docstring
+Returns the symbol used to render the item.
+
+.. seealso:: :py:func:`setSymbol`
+%End
+
+    void setSymbol( QgsLineSymbol *symbol /Transfer/ );
+%Docstring
+Sets the ``symbol`` used to render the marker item.
+
+The item takes ownership of the symbol.
+
+.. seealso:: :py:func:`symbol`
+%End
+
+  private:
+    QgsAnnotationLineStringItem( const QgsAnnotationLineStringItem &other );
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/annotations/qgsannotationlinestringitem.h                   *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
