+++ python/core/auto_generated/annotations/qgsannotationmarkeritem.sip.in
@@ -1,0 +1,70 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/annotations/qgsannotationmarkeritem.h                       *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+
+class QgsAnnotationMarkerItem : QgsAnnotationItem
+{
+%Docstring
+An annotation item which renders a marker symbol at a point location.
+
+.. versionadded:: 3.16
+%End
+
+%TypeHeaderCode
+#include "qgsannotationmarkeritem.h"
+%End
+  public:
+
+    QgsAnnotationMarkerItem( QgsPointXY point, const QgsCoordinateReferenceSystem &crs );
+%Docstring
+Constructor for QgsAnnotationMarkerItem, at the specified ``point`` in the given ``crs``.
+%End
+    ~QgsAnnotationMarkerItem();
+
+    virtual QString type() const;
+
+    virtual void render( QgsRenderContext &context, QgsFeedback *feedback );
+
+    virtual bool writeXml( QDomElement &element, QDomDocument &document, const QgsReadWriteContext &context ) const;
+
+    static QgsAnnotationMarkerItem *create() /Factory/;
+    virtual bool readXml( const QDomElement &element, const QgsReadWriteContext &context );
+
+    virtual QgsAnnotationMarkerItem *clone() /Factory/;
+
+
+    const QgsMarkerSymbol *symbol() const;
+%Docstring
+Returns the symbol used to render the marker item.
+
+.. seealso:: :py:func:`setSymbol`
+%End
+
+    void setSymbol( QgsMarkerSymbol *symbol /Transfer/ );
+%Docstring
+Sets the ``symbol`` used to render the marker item.
+
+The item takes ownership of the symbol.
+
+.. seealso:: :py:func:`symbol`
+%End
+
+  private:
+    QgsAnnotationMarkerItem( const QgsAnnotationMarkerItem &other );
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/annotations/qgsannotationmarkeritem.h                       *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
